test: move keyboard navigation tests to separate file
